[#391] Refactor notifications
