+++ api/handler/api.go
@@ -11,15 +11,22 @@ import (
 
 type (
 	handler struct {
-		log *zap.Logger
-		obj layer.Client
-		cfg *Config
+		log         *zap.Logger
+		obj         layer.Client
+		notificator Notificator
+		cfg         *Config
+	}
+
+	Notificator interface {
+		SendNotifications(topics map[string]string, p *SendNotificationParams) error
+		SendTestNotification(topic, bucketName, requestID, HostID string) error
 	}
 
 	// Config contains data which handler needs to keep.
 	Config struct {
-		DefaultPolicy *netmap.PlacementPolicy
-		DefaultMaxAge int
+		DefaultPolicy      *netmap.PlacementPolicy
+		DefaultMaxAge      int
+		NotificatorEnabled bool
 	}
 )
 
@@ -29,7 +36,7 @@ const DefaultPolicy = "REP 3"
 var _ api.Handler = (*handler)(nil)
 
 // New creates new api.Handler using given logger and client.
-func New(log *zap.Logger, obj layer.Client, cfg *Config) (api.Handler, error) {
+func New(log *zap.Logger, obj layer.Client, notificator Notificator, cfg *Config) (api.Handler, error) {
 	switch {
 	case obj == nil:
 		return nil, errors.New("empty NeoFS Object Layer")
@@ -37,9 +44,14 @@ func New(log *zap.Logger, obj layer.Client, cfg *Config) (api.Handler, error) {
 		return nil, errors.New("empty logger")
 	}
 
+	if cfg.NotificatorEnabled && notificator == nil {
+		return nil, errors.New("empty notificator")
+	}
+
 	return &handler{
-		log: log,
-		obj: obj,
-		cfg: cfg,
+		log:         log,
+		obj:         obj,
+		cfg:         cfg,
+		notificator: notificator,
 	}, nil
 }
