+++ api/handler/notifications_test.go
@@ -1,4 +1,4 @@
-package layer
+package handler
 
 import (
 	"testing"
